backend/requirements.txt
django
psycopg2-binary
django-oauth-toolkit
django-cors-headers
djangorestframework
gunicorn